Tip for Cache clearing when using API
Recently was trying to set a custom field after an entry was saved. It was fine on an update action, but on create, the Entry could not be found until I cleared the cache via `Stache::update()` which took a long time to find looking through the core files. I though this might help others in the same situation!